requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
Django==3.2.10
django-simpleui==2022.2.16
idna==3.3
importlib-metadata==4.8.3
lxml==4.7.1
Markdown==3.3.6
pendulum==2.1.2
Pillow==8.4.0
PyMySQL==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
pytzdata==2020.1
qiniu==7.5.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
